Components and localization were added
