+++ Localization/translations.js
@@ -1,14 +1,35 @@
 export const translations = {
-    'ua': {
+    'uk': {
       'Hello world!': 'Привіт світ!',
+      'Settings': 'Налаштування',
+      'Set language': 'Вибрати мову',
+      'Save': 'Зберегти',
+      'Error!': 'Помилка!',
+      'Please try again': 'Будь ласка, спробуйте ще раз',
+      'Success!': 'Успіх!',
+      'Settings saved': 'Налаштування збережено',
       'Now the performances premieres are :':'Зараз у прокаті'
     },
     'en': {
       'Hello world!': 'Hello world!',
+      'Settings': 'Settings',
+      'Set language': 'Set language',
+      'Save': 'Save',
+      'Error!': 'Error!',
+      'Please try again': 'Please try again',
+      'Success!': 'Success!',
+      'Settings saved': 'Settings saved',
       'Now the performances premieres are' : 'Now the performances premieres are'
     },
     'ru': {
       'Hello world!': 'Привет мир!',
+      'Settings': 'Настройки',
+      'Set language': 'Выбрать язык',
+      'Save': 'Сохранить',
+      'Error!': 'Ошибка!',
+      'Please try again': 'Пожалуйста, попробуйте еще раз',
+      'Success!': 'Успех!',
+      'Settings saved': 'Настройки сохранены',
       'Now the performances premieres are':'Сейчас в прокате :'
     },
   }
